# api/api.yaml
openapi: 3.0.0
info:
  title: Chat Application API
  description: API для мессенджера, включающего регистрацию, авторизацию, управление пользователями и диалогами.
  version: 1.0.0
paths:
  /auth/register:
    post:
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  example: "John"
                last_name:
                  type: string
                  example: "Doe"
                email:
                  type: string
                  example: "john.doe@example.com"
                nickname:
                  type: string
                  example: "johnny123"
                password:
                  type: string
                  example: "Password@123"
      responses:
        '201':
          description: Успешная регистрация
        '400':
          description: Ошибка валидации
  /auth/login:
    post:
      summary: Авторизация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "john.doe@example.com"
                password:
                  type: string
                  example: "Password@123"
      responses:
        '200':
          description: Успешная авторизация
        '401':
          description: Неверные учетные данные
...